Multiple mappings for a simple method is now properly implemented with retrocompatibility.
The 'method' attribute in a method can now be an object like the
following:

{
    "get" : "get URI",
    "put" : "put URI",
    "pull" : "pull URI",
    "delete" : "delete URI"
}

At least one of these must be defined. The keys in this object defines
which mappings are supported by the method. The value URI is the URI to
which the http is requested.

To mantain compatibility, method can still be a string, and
requestMapping must have the URI. Otherwise, requestMapping is ignored.

supportedMappings and postMapping are depricated and no longer
supported.

Work is left to be done on which URI is used to shown in the index page.