Now dynamic dates can be defined using %$d notation.
On a example getParam or within a json post file, a string may contain
%$d notation to generate dynamic dates. Notation goes as follows.

If a string contains %$d, then everything before should be interpretable
as date. If after %$d we have +1d it means we add 1 day to the date
defined before. -2d substracts to days. Same thing can be done for
months and years with, for example, +4m, -2y.

If %now is found after %$d, the modifications apply to the current date,
independently of what was before the separator.