Fix anchor hash location issue.
Issue description:
When we load any doc page with an anchor (i.e: #Books), and this anchor
is placed sufficiently below, at the end, when page is completely
loaded, we are placed in a wrong place. This is because the "toggle"
effect delay on the elements at loading time.

Solution:
At loading time we do not use this effect delay (which makes no sense
at all),but we keep this effect delay at click time.